Add memory region write support
This is the initial part required to support checkpoints that are stored
on the PC. This commit adds support for restoring memory regions.

In the `scripts` directory there is a python script called:
`restore-mem-serial.py`
This scripts works in tandem with the added code in main.c (have to move
it somewhere else).
It contains a small state machine to update multiple memory regions (see
the comments in the C and Python code).

The restoration data is stored in a .json file with the format:
{
    "restore": [
        {
          "addr_start": 20001838,
          "addr_end": 20001844,
          "data": [1,2,3,4,5,6,7,8,9,10,11,12]
        },
        ...
    ]
}

TODO:
 - move to a safe stack region before executing, otherwise we restore
 the stack we are working on. I already made an additional stack region
 in the linkerscript
 - add the ability to restore registers
 - add the checkpoint part :). This requires changing dump_meminfo in
 vm.c (a function I made to experiment with dumping the memory
 - update the python script to tie the check pointing and restoration
 together
 - ?
 - profit