Add python scripts to verify checkpoint restore
`gdb-memory-diff-dump.py`
is an python script that should be sourced by gdb. It places some
breakpoints and makes memory dumps of specific regions of memory. This
script will have to be changed when checkpointing regions etc are
changed but it serves as a base.

`memdiff.py`
is a python module or standalone program.
It is used to perform a binary diff and print the addresses that are
different between the two memory dumps.
When used by `gdb-memory-diff-dump.py` it also prints the variables that
have changed.

`gdbs-py`
is a script that starts gdb with the correct .elf file, connects to the
remote target (the Metro board) and also sources the python extension
(`gdb-memory-diff-dump.py`)

For the checkpoint procedure to work `scripts/checkpoint-serial.py`
should be running. As it is required to perform the checkpoint (it is
the non-volatile storage at the moment).