Disable garbage collector checkpoint
It is not working correctly over USB
When I try to restore the checkpoint the data send is probably to much
and the USB FIFO seems to get messed up.
Because we will move away from USB I don't want to spend time fixing it.

Because it's not in the .data or .bss section it is OK to ignore and act
as if it was restored correctly.
The configuration and GC state is saved and restored