Make the memory diff a bit less source line dependent
It now breaks on the restore point check in`checkpoint.c`
`gdb-memory-diff-dump.py` checks the actual value of
`checkpoint_svc_restore` to see if it was a checkpoint or restore.

On a checkpoint it overwrites the `after-cp` memory dump.
On a restore it overwrites the `after-restore` memory dump and compares
them.