Add NVM controller project code
The NVM memory and controller is based on the MSP430FR5994
microcontroller from TI. It has non-volatile FRAM memory.
The controller communication with CircuitPython is done over SPI.

The project is based on CMake. The `configure` script can be used for
the initial setup.
The project assumes that the MSP430 toolchain is installed in `/opt/`
For more details see `msp430-toolchain.cmake`

Some files in the project are not used but are there for legacy reasons.
These will be deleted at a later point.