Change GDB verification trigger
NB. Messy version that triggers before the register restore.
Will clean it up when the NVM is fully debugged, otherwise Hard faults
can happen before we diff the memory :(

Add the dummy function `gdb_break_me()` function.
The primary use of this function if to make it easy for GDB to break on.
Otherwise we need to specify a line number, which can be messy if the
source file changes, and a watchpoint on a variable will also trigger
when the .bss or .data section is initialized. This seemed to be the
easiest and most "portable" way.