Add safe_mode enum for USER_RESET
At default the safe_mode is set to USER_RESET instead of NO_SAFE_MODE
This means that when the system reboots because of the reset button it
will see this as RAM state is not lost.
Whenever a "fresh" boot is done (power was lost) the system will return
NO_SAFE_MODE.

When there is a USER_RESET CircuitPython will send a checkpoint_delete()
command to the NVM clearing existing checkpoitns.
Thus the reset button actually now does a full reset of the system and
power loss will cause the system to continue where it left off.