Add GDB command for power failure reboot
When we use 'mon_reset' it acts as a reset using the reset button. But
we want the reset button to also clear the NVM, and only reboots caused
by a power failure should trigger a restore.

By making the call to `port_set_saved_word()` to noinline we can call it
using GDB. In GDB when we want a reset that looks like it was caused by
a power failure we first call `port_set_saved_word(0)` as if the ram was
cleared (or messed up) by rebooting.