Checkpoint correct memory
Cicuitpython/Micropython uses their own allocator that allocates a stack
and a heap. The heap is used for the garbage collector.
The C stack is also included in the stack that is allocated using
`stack_init()`

I added the `allocations` table and the accompanying global variables to
the checkpoint content and modified the checkpoint sections in
checkpoint.c to go through the allocation table and checkpoint all the
entries.
Because this also includes the C stack I disabled it from the regions to
be checkpointed.

This commit also disables boot.py, it's not used by us so I don't
exactly know how it operates so it's disabled for now.