Add per byte synchronization
Sometimes the MSP is too slow to read the address values when they are
send in bulk. So now we synchronize these transactions per byte.

NB. This means that every NON-DMA transaction now is synchronized per
byte (by the MSP ready WR signal to the Metro).
If we observe that the DMA transaction speed is also affected we have a
bigger problem though. But that does not seem to be the case as of now.