[E-Paper Display EPD] Add a temperature eink lib
Adds a library to simply show the temperature on an e-ink

EPD interface code ported from:
https://github.com/wemos/LOLIN_EPD_Library

The EPD (E-Paper Display):
https://docs.wemos.cc/en/latest/d1_mini_shiled/epd_2_13.html

This library uses partial updates and serves as a demo application.
It can be extended to interface with either the traditional display
interface in circuitpython or a custom complete interface.

This demo only prints some assets to the screen and can show a 2-digit
number (the temperature).